Fix issue with ROOT-v6.16
In ROOT-v6.16, each RDataFrame node has shared ownership of its parent
and RLoopManager only has non-owning pointers to some of the nodes.
[ROOT-9467]

As a consequence, users do not have to take care that the head node of
the dataframe remains in the scope. However, PyRDF relied on this point,
always keeping at least one reference to the head node so intermediate
nodes of the graph had to be manually pruned in order to be deleted on
the cpp side. With the new ownership model in RDataFrame, if an
intermediate RDataFrame node, such as a Filter, does not have any
reference to it, it gets automatically deleted.

This commit solves this issue linking every new RDataFrame node created
on the cpp side with its corresponding PyRDF node, so the lifetime of
the cpp nodes depend on the lifetime of its equivalent PyRDF node.