Add tearDown method to restore current_backend
The setUpClass method in the ReducerMergeTest class selects the Spark backend
but it never disables it. This makes tests order dependent, since tests
that expect a default Local backend may fail when running after the
tests defined on this class.

This commit adds a tearDownClass method to restore the default backend
(Local) after running all tests.