Fix graph pruning
Add abstract class Proxy

The current Proxy class is aimed only at action nodes. It is a wrapper
around those nodes that triggers the execution of the computational graph
and retrieves the value of the action node, when the user issues an action
in the analysis. In order to tackle the graph pruning bug, the current idea
is to wrap all nodes (both transformations and actions) with proxies. With
this in mind, the current `Proxy` class will change name to `ActionProxy`
(for now) and a new abstract class will be created, taking the name of `Proxy`.

The `Proxy` abstract class is initialized with the attribute `node`, that is
the `PyRDF.Node` object linked when the user assigns a new variable to it.

When the user reassigns that same variable to another object (meaning that
he/she has no interest in the previous steps of analysis), the reference count
of the proxy decreases and the `__del__` method is called (automatically). This
abstract method (that will be implemented in both subclasses of the `Proxy`
class) has to flag the node that the current proxy was referring to as
prunable, effectively setting the `PyRDF.Node` attribute `prunable = True`.

Strictly speaking, in this commit no functionality of the project has been
modified. The `PyRDF.Node.prunable` attribute is set to `False` by default and
is only changed by action nodes that are not referenced by the user anymore,
effectively remaining silent. In a later commit this flag will be used to
enhance the `PyRDF.Node.graph_prune()` method.

Clarify names

Two class attributes previously introduced hadn't a clear meaning.

The attribute `node` in the `Proxy` class has been changed to `proxied_node`,
to better show the link between the proxy object and its wrapped node object.

The attribute `prunable` in the `Node` class was misleading. In fact, one could
think that if a node has `prunable == True`, then that node must be pruned from
the computational graph. This is not the case, since that attribute represents
the fact that a node is wrapped by a proxy which has a direct reference issued by
the user, e.g. via a variable assignment. As a consequence, when a user issues a
command such as `a = df.Filter().Define()`, only the proxy linked to the last
`Define()` node has a user reference, while the `Filter()` proxy doesn't so it
would flag the node as prunable even if it shouldn't be pruned.

For this reason, the attribute `prunable` has been changed to `has_user_references`
to better indicate its true meaning.

Add TransformationProxy, modify graph pruning logic

A class proxying transformation operations objects has been added, called
`TransformationProxy`, to go alongside with the `ActionProxy` class. Now all
nodes are always wrapped by proxy objects when they are created.

The `TransformationProxy` acts as its corresponding node. In fact, all
functions and attributes calls issued by the user on the proxy get mirrored
back to the node and return methods and attributes of the node itself. As for
the `ActionProxy`, also the `TransformationProxy` flags its wrapped node for
pruning when it gets garbage collected by Python (e.g. when the user is not
assigning a variable to it).

In order to stabilize the results of the graph pruning function, it has been split
in two separate methods. The new method `is_prunable` checks whether a node
can be pruned from the graph. The first condition is that the node has no
children. If this holds, then the attribute `has_user_references` is checked
together with the condition that the node is an action node already computed.
If either of these is `False`, then the method returns `True`. The `graph_prune`
method already present now recurses through the graph, starting from the node
on which is called and evaluating the `is_prunable` method on all of the
children. If a child is prunable, it will not be added to the children list
of the node, otherwise it will.

Other changes:

The main changes of this commit reflect also on some tests, that had to be
changed accordingly. Wherever there was a call to a transformation operation
through a variable, which previously was of class `PyRDF.Node` and now of class
`PyRDF.TransformationProxy`, it has been changed to the attribute `proxied_node`
of the proxy, in order to avoid unmeaningful errors.

Furthermore, it has been noticed that pickling and unpickling a proxy object
through direct call of the functions `pickle.dumps()` / `pickle.loads()`, throws
a `RecursionError` since pickle is not able to reconstruct the link between
proxy and object. This is not considered an error, since our interest is in
pickling the computational graph objects directly without the proxies. In future,
if needed, this could be patched by redefining the `__getstate__` and `__setstate__`
magic methods of the proxy object.

Finally, in `test_dist.py` the function `rangesToTuples` makes use of the
Python `map()` function, which in Python 2 returns a list, while in Python 3
returns an iterator. This behaviour has been fixed by wrapping the returned value
with `list()`.

Add proxy tests

The newly created proxy classes need to be tested. The new tests check the
following features:

- The abstract class `Proxy` cannot be instantiated by the user
- All subclasses of `Proxy` cannot be instantiated without defining
  the `__del__` abstract method
- The variable assigned to a proxy is of the correct type
- The `proxied_node` attribute of a proxy is of the correct type
- The TransformationProxy correctly reads attributes passed to it,
  recognizes if they are supported operations or other attributes
  and returns the corresponding methods or attributes of the proxied
  node
- The `has_user_references` attribute of a node proxied by a
  `TransformationProxy` object is equal to `True` until the proxy is
  assigned to a variable
- The function `__getstate__` of a `TransformationProxy` object is
  called when the proxy is pickled. This is to check the behaviour
  of pickle, knowing that right now a proxy cannot be unpickled. In
  order to track the calls to the function a decorator has been added
  to the `__getstate__` method of the `TransformationProxy` class.

Fix graph pruning