Add documentation, change tests
In order to reflect the change of logic of the last commit, some
docstrings have been added to `RDataFrame` class and to the renewed
proxy family of classes.

Furthermore, many tests have been modified in order to account for the
fact that now all attributes and methods of nodes are accessed through
their proxies.

A minor issue has been noticed, that in Python 3 abstract classes that
do not implement abstract methods can be instantiated anyway, so for the
moment there is no need to test that the `Proxy` class cannot be
instantiated. This will be addressed in later commits.