Clear division of tasks in proxy subclasses
The tasks that the family of proxy classes should manage now are more clearly assigned
to either the `TransformationProxy` or `ActionProxy` subclasses.

All nodes that make changes to data (The head node that collects new data, the
transformations that modify their structure) are proxied by `TransformationProxy`s.
Only nodes that hold action operations are proxied by `ActionProxy`s.

The first subclass handles checks on the attribute called on the node. If it is a node
attribute, the proxy will return its value, or if it is an operation name it will return
the appropriate proxy for that operation, after creating the node and the operation
objects linked to it.

The second subclass keeps its original behaviour, attribute calls return the value
of the action, or trigger the execution of the graph if it has not been done yet.

The `Proxy` parent class is now an abstract class with abstract method `__getattr__`.