include header directories
The user can now write a string containing the path of a directory
with c++ headers as the `PyRDF.include()` argument. If the string
is an existing path, the function will traverse all files in the
directory and will create a list with the relative paths of the
headers. The elements of the list will then be added to the
`include` statements of the c++ code.