create new head node when new graph is issued
In the local backend, the execution of the RDataFrame graph is started
with the head node of the graph as parameter (`pyroot_rdf`). Before that,
the `execute` method of the `Local` backend class checks whether an
RDataFrame already exists. If not, it creates a new one.

When the user creates a new RDF in the same script, the execution of
the graph will begin with the first RDF as head node, not the last one
created by the user.

This is solved via double checking that the existing `pyroot_rdf` is the
same as the head node of the current graph. If they are different, a new
RDF is created and used as the starting node of the graph.