extend include to shared libraries (and pcm files)
The user can now input any combination of single path or list of paths
to a shared library or a directory containing shared libraries.

The new functions `include_shared_libraries` and `Utils.declare_shared_libraries`
will handle the retrieval and declaration of the shared libraries input
by the user. Caveat: the shared libraries will be included in the analysis
directly via `ROOT.gSystem.Load()`, the user has to handle the inclusion of
the headers related to the shared libraries via the normal `include` function
or declare all functions and objects implemented in the libraries that s/he
needs through calls to `ROOT.gInterpeter.ProcessLine()`.

The new test to check the new feature creates a shared library from a `.cpp`
file at runtime, performs the analysis and finally removes the `.so` file
from the system to avoid the upload.

Since this test uses many long strings for storing the paths to the various
files, it throws some PEP8 warnings. For the moment these are suppressed,
to be seen.

The `PyRDF.include_shared_libraries()` function will try to automatically retrieve
pcm files in the same folder as the shared libraries, in order to send them to the
distributed workers. pcm files must be in the same folder as their corresponding
shared libraries and they only have to be distributed, not declared.

A function `send_pcm_files` is provided if the user for some reason has the pcm
files in a separate directory, just to send those to the workers.

A new test creates a shared library with its pcm file, checks that ROOT correctly
read it and performs some simple operations on an RDataFrame.