[WIP] send headers to executors and declare them at runtime
Right now, when executing PyRDF on a distributed backend (such as Spark), if the user
has included any headers, they will just be declared on the local machine (spark driver node).

In order to send the workers all the functions needed for the analysis, Spark provides a specific
pattern through `SparkContext.addFile()` and `SparkFiles.get()`.

* `addFile()` can be called before the analysis (MapReduce job) has started, to send whatever
file to the driver node and then to the workers. It receives in input a string with the path
(relative to the current python session) to the file(s) that have to be sent to the worker(s).
It is a method of the `SparkContext` class, so it can be called only if an instance is currently
active on the driver node.
* `SparkFiles.get()` must be called during the Map phase (which is executed on the workers) to
retrieve the path(s) of the file(s) on the specific worker (since each worker will have different path(s)
to the file(s) ). It receives in input a string with the filename only (not the path)

The new implementation calls `addFile()` inside `PyRDF.include()` and `SparkFiles.get()` inside the
`map` function of the `Dist` class.

A big issue is that the retrieval of the paths to the headers is specific of the backend, so there should
be some `get_distributed_files` function living inside each backend and performing different things in
order to retrieve the headers. Unfortunately, an error

"It appears that you are attempting to reference SparkContext from a broadcast"

occurs when trying to call the `get_distributed_files` from the `Spark` subclass inside the `map`
method of `Dist` parent class. Further investigation needed