start sphinx docs with travis integration
A new documentation project with Sphinx has been added to the repo, under
the docs/ folder. The .travis.yml file has been updated to also include
building the docs and publishing them to the gh-pages branch, in order
to be published on github pages. Settings of the repository have been changed accordingly.