Fix fallback to local execution when data source has zero entries
The `Dist` backend checks that the file from which the RDataFrame was instantiated has
more than zero entries. If this is not the case, the execution of the graph is sent
to the local backend via a direct call to `Local.execute()` method.

Previously, this call was to the class, not the instance, so it was missing the `self` parameter.
Now, the `current_backend` variable is imported and changed to `Local()`. Then the call
to `execute` is done on the variable.

To go along with this change, a new test has been added that gets the entries of an histogram
made from an empty .root file