Change lists to sets in include functions
All the `include_*` functions work with lists and change global variables that
are lists too. All of these objects are meant to store paths to headers, shared libraries
and all other files that are needed in the analysis. As such, they should not contain duplicates.

Python lists allow for duplicate items, so another data structure that takes care of this is needed.
In this pr, lists are converted to sets, that are unordered collections of items and do not allow
duplicates.

A minor concern would be that in some cases (badly written) C++ programs rely on the order in which
the headers are included. But generally this should be the case and the users should be advised
against such behaviours.

The test for the functionalities of the functions changed has been tweaked a bit to accomodate for
the fact that now there are unordered collections instead of ordered lists. In this sense, unittest
provides the function `assertCountEqual` in Python 3 and `assertItemsEqual` in Python 2. Thus, a
feature detection has been implemented in the test to distinguish between the two versions.