Start documentation building on Read The Docs (#68)
Read The Docs (rtd) is an established website for hosting technical documentation.
This PR creates the basic configuration files to upload and synchronize
PyRDF docs with an rtd account. The  packages that allow building a beautiful
documentation layout have been added as requirements in `docs/requirements.txt`

Caveat: Right now, the docs are built with `sphinx.automodule` that executes
all the code while parsing for docstrings. This means that docs have to be built
inside a fully-functional environment (e.g. the one created with the usual testing configuration file).
On the rtd website there are strict limitations for time and cpu usage, so a special
request was needed since conda takes a while to setup.