Change npartitions attribute of Spark class instance (#72)
* Change npartitions attribute of Spark class instance

The `npartitions` attribute of the `Spark` class previously didn't get changed when number of partitions typed by the user was greater than number of clusters in the ROOT file.
Now both the internal variable and the class attribute change, so that both the driver and the workers are synchronized on the number of partitions of the dataset.

Since now the attribute is changed during the creation of the ranges to divide the RDataFrame in, it has been moved to the `Dist` parent class. This should be expected since any distributed backend will probably leverage some kind of data partitioning before starting the job.
In any case, extra modifications to the `npartitions` attribute are still done in the `Spark` subclass as before.

A test has also been added to check that the class attributed is correctly changed. This test is encapsulated in a new `unittest.TestCase` class because somehow was interfering with other tests with the `Spark` backend instantiated.
Further investigation is needed to understand the behaviour of the `unittest` tests context and how it is managed.

* npartitions is always used as class attribute