Refactor Travis CI configuration to work with ROOT nightlies
Recently a conda package of the ROOT nightlies has been added at https://root.cern/download/conda-nightly/latest/
This is very useful to test PyRDF with the latest changes in ROOT. Notably in this moment the RMergeableValue classes are scheduled to be released with 6.24 while the latest release is 6.22

The nightlies package requires python3.8, this means that we would be losing testing against python 2 . I do not consider this a negative since python2 is no longer supported.
Anyway, come ROOT 6.24 it will still be possible to reintegrate python 2 CI through the updated conda package.