Refactor Travis CI configuration to work with ROOT nightlies (#101)
Recently a conda package of the ROOT nightlies has been added at https://root.cern/download/conda-nightly/latest/
This is very useful to test PyRDF with the latest changes in ROOT. Notably in this moment the RMergeableValue classes are scheduled to be released with 6.24 while the latest release is 6.22

A new conda env file has been added with python 3.8 as default version that will run with ROOT nightlies. The python 2.7 and 3.7 environments will run with the latest release. The code style check step of the CI has been updated to use python 3.7